Changes in Sandbox demo change: Loaded some more pictures for the demo.
feature: Image::Load now has an overload that takes in a width and height,
and uses these values to set the width and height of the loaded texture.

fix: EventListenerSystem::OnEvent now checks if the entity has a
RectComponent and a Coordinate2DComponent, and if not, the function simply
returns.